requirements.txt
face_recognition
opencv-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
coloredlogs
